levels/level_01_prompt_only/pyproject.toml
# Project-wide definition and dependency management via TOML.
# Follows PEP 518 / PEP 621 — the structure and file name are standardized.
# Main purpose: declare dependencies, build settings, and tool configuration.
# Assumes uv as the package manager.
[project]
name = "level_01_prompt_only"
version = "0.1.0"
description = "Minimal chatbot MVP — prompt only, no memory, no external data"
requires-python = ">=3.11"
dependencies = [
    "openai",
    "pydantic",
    "pyyaml"
]